The testing task
This is where I add the site files. I used BEM methodology, scss with prepros program and flexbox. Due to the fact that there was no menu, I did not add it to the burger. I also used the author's development of automatic font change depending on the screen size, developed on scss.